struct: Move templates to separate package(dir)
