+++ binaryTree/src/main/kotlin/org/tree/binaryTree/Nodes.kt
@@ -1,5 +1,7 @@
 package org.tree.binaryTree
 
+import org.tree.binaryTree.templates.TemplateNode
+
 class Node<T : Comparable<T>>(v: T) : TemplateNode<T, Node<T>>(v)
 
 class RBNode<T : Comparable<T>>(p: RBNode<T>?, v: T) : TemplateNode<T, RBNode<T>>(v) {
