+++ binaryTree/src/main/kotlin/org/tree/binaryTree/templates/TemplateBSTree.kt
@@ -1,4 +1,4 @@
-package org.tree.binaryTree
+package org.tree.binaryTree.templates
 
 abstract class TemplateBSTree<T : Comparable<T>, NODE_T : TemplateNode<T, NODE_T>> {
     var root: NODE_T? = null
