+++ binaryTree/src/main/kotlin/org/tree/binaryTree/trees/AVLTree.kt
@@ -1,7 +1,7 @@
 package org.tree.binaryTree.trees
 
 import org.tree.binaryTree.AVLNode
-import org.tree.binaryTree.TemplateBalanceBSTree
+import org.tree.binaryTree.templates.TemplateBalanceBSTree
 import kotlin.math.max
 
 class AVLTree<T : Comparable<T>> : TemplateBalanceBSTree<T, AVLNode<T>>() {
