+++ binaryTree/src/main/kotlin/org/tree/binaryTree/trees/BinSearchTree.kt
@@ -1,7 +1,7 @@
 package org.tree.binaryTree.trees
 
 import org.tree.binaryTree.Node
-import org.tree.binaryTree.TemplateBSTree
+import org.tree.binaryTree.templates.TemplateBSTree
 
 class BinSearchTree<T : Comparable<T>> : TemplateBSTree<T, Node<T>>() {
     override fun insert(curNode: Node<T>?, obj: T): Node<T>? {
